refactor: UI 텍스트 수정 등 - 하드코딩 UI 텍스트를 리소스 값(strings.xml)으로 수정 - 프래그먼트 간 정보전달을 위한 동반객체 위치 및 네이밍 수정
